requirements.txt

flask==2.3.3
langchain==0.2.14
langchain-community==0.2.12
langchain-chroma==0.1.2
langchain-google-genai==1.0.8
pypdf==4.3.1
chromadb==0.5.3
google-generativeai==0.7.2
gunicorn==22.0.0
langchain-core
werkzeug